+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## ChangeLog
+#### Version 0.8.2 (not yet released)
+- [***change:***] Renamed `absolute:` to `file:`
+- [***change:***] Removed `relative:` attachment prefix
+- [enhancement:] New `res:` attachment prefix
+- [enhancement:] New `www:` attachment prefix
+
 #### Version 0.8.1 (06.04.2014)
 - [enhancement:] Make use Cordovas NSData+Base64 extension.
 - [enhancement:] Log error message if attachment path does not exist.